Compile Fixes.

+++ signal-service/src/main/java/com/algomeet/signalservice/dto/OneTimePreKeyResponse.java
@@ -12,7 +12,7 @@ public class OneTimePreKeyResponse {
 	private long id;
 	private UUID userKey;
 	private Integer deviceId;
-	private Integer preKeyId;
+	private Long preKeyId;
 	private String publicKey;
 	private Boolean used;
 	private Instant createdAt;
